requirements.txt
Flask
python-dotenv
cx_Oracle
